Fixes #208 - Freeseer fails to record with VideoPreview and File output plugins
When using the VideoPreview plugin along with recording to a File plugin at the
same time Freeseer will fail to record and produce an empty file.

This fix adds an additional ffmpegcolorspace to the gstreamer pipeline by
placing it inside of the VideoPreview plugin which seems to resolve this issue.